Allow the Passenger restart method to be configured.
Previously it was hardcoded to use the `passenger-config` program if the
Passenger version is greater than 4. However, this program can have
issues on certain systems such as being unable to access the /tmp dir
with the PrivateTmp systemd configuration[1] or requiring a privileged
user[2].

The behaviour has the same defaults as before but can be overridden by
setting the `:passenger_restart_method` setting to `:restart_txt` to use
the method before Passenger 5. Touching restart.txt is still supported
but with a 10 second (by default) polling interval.

[1]: https://code.google.com/p/phusion-passenger/issues/detail?id=1106
[2]: https://github.com/phusion/passenger/issues/1392