Fix `passenger-config` not being run with 'cd #{release_path}' when passenger_in_gemfile is true
The old method preprocesses the command with SSHKit::Command.new
in order to allow the 'passenger-config' command to be transformed
via the SSHKit command map, even when the entire command is prepended
with 'sudo'. However, that approach broke cases in which
passenger_restart_with_sudo is false, because a string (instead of a
command symbol) is passed to `execute`. As a result, `execute` does
not prepend `cd #{release_path}` to the final command. This would
normally not cause problems, but would cause problems when passenger_in_gemfile
is true because 'bundle exec passenger-config' tries to look for a Gemfile
in the current working directory.